+++ projects/journey-maps-client-testapp/src/app/app.component.ts
@@ -24,6 +24,7 @@ export class AppComponent {
   // Initial map position
   zoomLevel = 7.5;
   mapCenter: LngLatLike = [7.4391326448171196, 46.948834547463086];
+  selectedMarker: Marker;
 
   // Can be used instead of zoomLevel and mapCenter
   boundingBox: LngLatBoundsLike = [[-9.97708574059, 51.6693012559], [-6.03298539878, 55.1316222195]];
@@ -39,6 +40,7 @@ export class AppComponent {
       subtitle: 'Rent a Bike - Ihr Mietvelo',
       position: [7.5897, 47.5476],
       category: MarkerCategory.INFORMATION,
+      emitOnSelect: true,
       infoBlocks: [
         this.infoBlockFactoryService.createTextInfoBlock(
           'Verfügbare Velotypen',
@@ -65,6 +67,7 @@ export class AppComponent {
       subtitle: 'My home is my castle',
       position: [7.296515, 47.069815],
       category: MarkerCategory.INFORMATION,
+      emitOnSelect: true,
       infoBlocks: [
         this.infoBlockFactoryService.createTextInfoBlock(
           this.loremIpsum.generateWords(3),
@@ -92,6 +95,7 @@ export class AppComponent {
       category: MarkerCategory.CUSTOM,
       icon: 'assets/icons/train.png',
       iconSelected: 'assets/icons/train_selected.png',
+      emitOnSelect: true,
       infoBlocks: [
         this.infoBlockFactoryService.createHtmlInfoBlock(
           'Cupcake Ipsum',
@@ -109,6 +113,7 @@ export class AppComponent {
       subtitle: 'SBB Wylerpark',
       position: [7.446450, 46.961409],
       category: MarkerCategory.WARNING,
+      emitOnSelect: true,
       infoBlocks: [
         this.infoBlockFactoryService.createButtonInfoBlock(
           'Show menu plan',
@@ -117,4 +122,8 @@ export class AppComponent {
       ]
     },
   ];
+
+  onSelectedMarker(event): void {
+    this.selectedMarker = event;
+  }
 }